+++ gwnr/waveform/__init__.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from .align import *
 from .eccentric import *
 
-from . import enigma_utils
+from . import esigma_utils
 from .hybridize import *
 from .nr_waveform_sxs import *
 from .tidal import *
