+++ gwnr/waveform/esigma_utils.py
